Move app views into modules
