+++ indigo_app/views/works.py
@@ -3,10 +3,7 @@ import json
 import io
 import re
 
-from django.conf import settings
 from django.core.exceptions import ValidationError
-from django.contrib.auth.mixins import PermissionRequiredMixin
-from django.contrib.auth.views import redirect_to_login
 from django.contrib import messages
 from django.views.generic import DetailView, TemplateView, FormView, UpdateView, CreateView
 from django.views.generic.list import MultipleObjectMixin
@@ -16,57 +13,17 @@ from django.shortcuts import redirect, get_object_or_404
 from reversion import revisions as reversion
 from cobalt.act import FrbrUri
 from datetime import datetime
-from allauth.utils import get_request_param
 import requests
 import unicodecsv as csv
 
-from indigo_api.models import Document, Subtype, Work, Amendment
-from indigo_api.serializers import DocumentSerializer, WorkSerializer, WorkAmendmentSerializer
+from indigo_api.models import Document, Subtype, Work, Amendment, Country
+from indigo_api.serializers import WorkSerializer, DocumentSerializer
 from indigo_api.views.documents import DocumentViewSet
 from indigo_api.signals import work_changed
-from indigo_api.models import Language, Country
 from indigo_app.revisions import decorate_versions
-from indigo_app.forms import DocumentForm, UserForm, BatchCreateWorkForm
+from indigo_app.forms import DocumentForm, BatchCreateWorkForm
 
-
-class IndigoJSViewMixin(object):
-    """ View that inject's the appropriate Backbone view name into the template, for use by the Backbone
-    view system. By default, the Backbone view name is the same name as the view's class name.
-    Set `js_view` to another name to change it, or the empty string to disable completely.
-    """
-    js_view = None
-
-    def get_context_data(self, **kwargs):
-        context = super(IndigoJSViewMixin, self).get_context_data(**kwargs)
-        context['js_view'] = self.get_js_view()
-        return context
-
-    def get_js_view(self):
-        if self.js_view is None:
-            return self.__class__.__name__
-        return self.js_view
-
-
-class AbstractAuthedIndigoView(PermissionRequiredMixin, IndigoJSViewMixin):
-    """ Abstract view for authenticated Indigo views.
-    """
-    # permissions
-    raise_exception = True
-    permission_required = ()
-    must_accept_terms = True
-
-    def dispatch(self, request, *args, **kwargs):
-        if not request.user.is_authenticated:
-            return redirect_to_login(self.request.get_full_path(), self.get_login_url(), self.get_redirect_field_name())
-
-        if not self.has_permission():
-            return self.handle_no_permission()
-
-        if self.must_accept_terms and not request.user.editor.accepted_terms:
-            # user must accept terms
-            return redirect_to_login(self.request.get_full_path(), 'accept_terms', self.get_redirect_field_name())
-
-        return super(AbstractAuthedIndigoView, self).dispatch(request, *args, **kwargs)
+from .base import AbstractAuthedIndigoView
 
 
 class LibraryView(AbstractAuthedIndigoView, TemplateView):
@@ -496,97 +453,3 @@ class ImportDocumentView(AbstractWorkDetailView):
         context['language'] = self.request.GET.get('language') or work.country.primary_language.code
 
         return context
-
-
-class DocumentDetailView(AbstractAuthedIndigoView, DetailView):
-    model = Document
-    context_object_name = 'document'
-    pk_url_kwarg = 'doc_id'
-    template_name = 'document/show.html'
-
-    # permissions
-    permission_required = ('indigo_api.view_work',)
-
-    def get_object(self, queryset=None):
-        doc = super(DocumentDetailView, self).get_object(queryset)
-        if doc.deleted:
-            raise Http404()
-        return doc
-
-    def get_context_data(self, **kwargs):
-        context = super(DocumentDetailView, self).get_context_data(**kwargs)
-
-        doc = self.object
-
-        context['work'] = doc.work
-        context['work_json'] = json.dumps(WorkSerializer(instance=doc.work, context={'request': self.request}).data)
-        context['country'] = doc.work.country
-        context['locality'] = context['country'].work_locality(doc.work)
-
-        # TODO do this in a better place
-        context['countries'] = Country.objects.select_related('country').prefetch_related('locality_set', 'publication_set', 'country').all()
-        context['countries_json'] = json.dumps({c.code: c.as_json() for c in context['countries']})
-
-        context['document_content_json'] = json.dumps(doc.document_xml)
-
-        serializer = WorkSerializer(context={'request': self.request}, many=True)
-        works = Work.objects.filter(country=doc.work.country)
-        context['works_json'] = json.dumps(serializer.to_representation(works))
-
-        context['amendments_json'] = json.dumps(
-            WorkAmendmentSerializer(context={'request': self.request}, many=True)
-            .to_representation(doc.work.amendments))
-
-        context['form'] = DocumentForm(instance=doc)
-        context['countries'] = Country.objects.select_related('country').prefetch_related('locality_set', 'publication_set', 'country').all()
-        context['countries_json'] = json.dumps({c.code: c.as_json() for c in context['countries']})
-        context['subtypes'] = Subtype.objects.order_by('name').all()
-        context['languages'] = Language.objects.select_related('language').all()
-
-        serializer = DocumentSerializer(context={'request': self.request}, many=True)
-        context['documents_json'] = json.dumps(serializer.to_representation(DocumentViewSet.queryset.all()))
-        return context
-
-
-class EditAccountView(AbstractAuthedIndigoView, FormView):
-    template_name = 'indigo_app/user_account/edit.html'
-    form_class = UserForm
-
-    def get_success_url(self):
-        return reverse('edit_account')
-
-    def get_form_kwargs(self):
-        kwargs = super(EditAccountView, self).get_form_kwargs()
-        kwargs['instance'] = self.request.user
-        return kwargs
-
-    def form_valid(self, form):
-        self.object = form.save()
-        return super(EditAccountView, self).form_valid(form)
-
-
-class EditAccountAPIView(AbstractAuthedIndigoView, DetailView):
-    context_object_name = 'user'
-    template_name = 'indigo_app/user_account/api.html'
-
-    def get_object(self):
-        return self.request.user
-
-    def post(self, request):
-        request.user.editor.api_token().delete()
-        # force a new one to be created
-        request.user.editor.api_token()
-        return self.get(request)
-
-
-class AcceptTermsView(AbstractAuthedIndigoView, UpdateView):
-    context_object_name = 'editor'
-    template_name = 'indigo_app/user_account/accept_terms.html'
-    fields = ('accepted_terms',)
-    must_accept_terms = False
-
-    def get_object(self):
-        return self.request.user.editor
-
-    def get_success_url(self):
-        return get_request_param(self.request, self.get_redirect_field_name(), settings.LOGIN_REDIRECT_URL)
